+++ hutool-core/src/main/java/cn/hutool/core/date/DatePattern.java
@@ -17,13 +17,14 @@ public class DatePattern {
 	/**
 	 * 标准日期时间正则，每个字段支持单个数字或2个数字，包括：
 	 * <pre>
+	 *     yyyy-MM-dd HH:mm:ss.SSS
 	 *     yyyy-MM-dd HH:mm:ss
 	 *     yyyy-MM-dd HH:mm
 	 *     yyyy-MM-dd
 	 * </pre>
 	 * @since 5.3.6
 	 */
-	public static final Pattern REGEX_NORM = Pattern.compile("\\d{4}-\\d{1,2}-\\d{1,2}( \\d{1,2}:\\d{1,2}(:\\d{1,2})?)?");
+	public static final Pattern REGEX_NORM = Pattern.compile("\\d{4}-\\d{1,2}-\\d{1,2}(\\s\\d{1,2}:\\d{1,2}(:\\d{1,2})?)?(.\\d{1,3})?");
 
 	//-------------------------------------------------------------------------------------------------------------------------------- Normal
 	/** 标准日期格式：yyyy-MM-dd */