+++ hutool-core/src/test/java/cn/hutool/core/date/DateUtilTest.java
@@ -779,6 +779,9 @@ public class DateUtilTest {
 		Assert.assertEquals("2020-05-08 03:02:03", dateTime.toString());
 		dateTime = DateUtil.parse("2020-5-8 3:12:13");
 		Assert.assertEquals("2020-05-08 03:12:13", dateTime.toString());
+
+		dateTime = DateUtil.parse("2020-5-8 4:12:26.223");
+		Assert.assertEquals("2020-05-08 04:12:26", dateTime.toString());
 	}
 
 	@SuppressWarnings("ConstantConditions")