+++ hutool-db/src/main/java/cn/hutool/db/ds/pooled/PooledDSFactory.java
@@ -1,9 +1,10 @@
 package cn.hutool.db.ds.pooled;
 
-import javax.sql.DataSource;
-
+import cn.hutool.core.util.StrUtil;
 import cn.hutool.db.ds.AbstractDSFactory;
 import cn.hutool.setting.Setting;
+
+import javax.sql.DataSource;
 
 /**
  * Hutool自身实现的池化数据源工厂类
@@ -37,6 +38,15 @@ public class PooledDSFactory extends AbstractDSFactory {
 		dbConfig.setMinIdle(poolSetting.getInt("minIdle", 0));
 		dbConfig.setMaxActive(poolSetting.getInt("maxActive", 8));
 		dbConfig.setMaxWait(poolSetting.getLong("maxWait", 6000L));
+
+		// remarks等特殊配置，since 5.3.8
+		String connValue;
+		for (String key : KEY_CONN_PROPS) {
+			connValue = poolSetting.get(key);
+			if(StrUtil.isNotBlank(connValue)){
+				dbConfig.addConnProps(key, connValue);
+			}
+		}
 
 		return new PooledDataSource(dbConfig);
 	}