+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-## 5.3.8 (2020-06-15)
+## 5.3.8 (2020-06-16)
 
 ### 新特性
 * 【core   】     增加ISO8601日期格式（issue#904@Github）
@@ -12,6 +12,8 @@
 * 【core   】     复制创建一个Bean对象, 并忽略某些属性(pr#130@Gitee)
 * 【core   】     DateUtil.parse支持更多日期格式(issue#I1KHTB@Gitee)
 * 【crypto 】     增加获取密钥空指针的检查(issue#925@Github)
+* 【core   】     增加StrUtil.removeAny方法(issue#923@Github)
+* 【db     】     增加部分Connection参数支持(issue#924@Github)
 
 ### Bug修复
 * 【json   】     修复append方法导致的JSONConfig传递失效问题（issue#906@Github）